Clean up
Put all classes in different files and organized the files.
