+++ desktop_app/UI/app_manager.py
@@ -1,0 +1,6 @@
+# This class will manage all interaction with the different scripts.
+
+class App_Manager():
+
+    def print_out_things(thing):
+        print(thing)
