--- desktop_app/main_manager.py
@@ -1,7 +1,0 @@
-
-
-# This class will manage all interaction with the different scripts.
-
-
-def print_out_things(thing):
-    print(thing)
